+++ common/internal-types/Cargo.toml
@@ -13,16 +13,10 @@ crate-type = ["rlib"]
 
 [features]
 default = []
-serde = [
-  "dep:serde",
-  "dep:serde_bytes",
-  "hopr-crypto-types/serde",
-  "bloomfilter/serde",
-]
+serde = ["dep:serde", "dep:serde_bytes", "hopr-crypto-types/serde"]
 
 [dependencies]
 aquamarine = { workspace = true }
-bloomfilter = { workspace = true }
 hex-literal = { workspace = true }
 hex = { workspace = true }
 num_enum = { workspace = true }
@@ -46,7 +40,3 @@ lazy_static = { workspace = true }
 
 [package.metadata.cargo-machete]
 ignored = ["serde_bytes"]
-
-[[bench]]
-name = "bloom_filter"
-harness = false
