+++ transport/bloom/benches/bloom_filter.rs
@@ -2,7 +2,7 @@ use std::hint::black_box;
 
 use criterion::{Criterion, Throughput, criterion_group, criterion_main};
 use hopr_crypto_types::types::PACKET_TAG_LENGTH;
-use hopr_internal_types::prelude::TagBloomFilter;
+use hopr_transport_bloom::TagBloomFilter;
 
 fn tag_bloom_filter_bench(c: &mut Criterion) {
     // Fill up the Bloom filter first
