+++ transport/protocol/Cargo.toml
@@ -15,8 +15,6 @@ prometheus = ["dep:hopr-metrics", "hopr-path/prometheus"]
 
 [dependencies]
 async-trait = { workspace = true }
-async-lock = { workspace = true }
-bincode = { workspace = true }
 futures = { workspace = true }
 hex-literal = { workspace = true }
 lazy_static = { workspace = true }
@@ -46,6 +44,7 @@ hopr-metrics = { optional = true, workspace = true }
 hopr-path = { workspace = true }
 hopr-platform = { workspace = true }
 hopr-primitive-types = { workspace = true }
+hopr-transport-bloom = { workspace = true, features = ["persistent"] }
 hopr-transport-identity = { workspace = true }
 hopr-transport-packet = { workspace = true }
